Extract shared ConfettiSearch component for search views
Refactored search bar implementation to reduce duplication:
- Created ConfettiSearch inside the shared components package.
- Replaced custom topBar search row layouts in SearchView and ConferenceListView with ConfettiSearch.
- Cleaned unused properties and custom key interceptors from SearchView.

Test: ./gradlew :androidApp:testDebugUnitTest
